backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy[asyncio]==2.0.48
aiosqlite==0.22.1
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
structlog==25.5.0
alembic==1.18.4
anthropic==0.84.0
sentence-transformers==4.1.0
faiss-cpu==1.11.0
cachetools==5.5.2
